Introduce `Placeholder`, rename `ErrorPlaceholder` to `StandardPlaceholder`
