+++ context/src/main/kotlin/io/spine/tools/validation/option/required/RequiredOption.kt
@@ -64,9 +64,9 @@ import io.spine.tools.validation.option.IF_MISSING
 import io.spine.tools.validation.option.REQUIRED
 import io.spine.tools.validation.option.checkPrimaryApplied
 import io.spine.tools.validation.option.required.RequiredFieldSupport.isSupported
-import io.spine.validation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.StandardPlaceholder.FIELD_PATH
+import io.spine.validation.StandardPlaceholder.FIELD_TYPE
+import io.spine.validation.StandardPlaceholder.PARENT_TYPE
 
 /**
  * Controls whether a field should be validated as `(required)`.
