+++ java/src/main/kotlin/io/spine/tools/validation/java/generate/option/ChoiceGenerator.kt
@@ -47,9 +47,9 @@ import io.spine.tools.validation.java.generate.ValidateScope.parentPath
 import io.spine.tools.validation.java.generate.ValidateScope.violations
 import io.spine.tools.validation.option.CHOICE
 import io.spine.validation.ConstraintViolation
-import io.spine.validation.ErrorPlaceholder
-import io.spine.validation.ErrorPlaceholder.GROUP_PATH
-import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.Placeholder
+import io.spine.validation.StandardPlaceholder.GROUP_PATH
+import io.spine.validation.StandardPlaceholder.PARENT_TYPE
 
 /**
  * The generator for the `(choice)` option.
@@ -110,7 +110,7 @@ private class GenerateChoice(private val view: ChoiceOneof) {
     private fun supportedPlaceholders(
         groupPath: Expression<FieldPath>,
         typeName: Expression<String>,
-    ): Map<ErrorPlaceholder, Expression<String>> = mapOf(
+    ): Map<Placeholder, Expression<String>> = mapOf(
         GROUP_PATH to groupPath.joinToString(),
         PARENT_TYPE to typeName
     )
