+++ java/src/main/kotlin/io/spine/tools/validation/java/generate/option/DistinctGenerator.kt
@@ -61,12 +61,12 @@ import io.spine.tools.validation.java.generate.ValidateScope.parentPath
 import io.spine.tools.validation.java.generate.ValidateScope.violations
 import io.spine.tools.validation.option.PATTERN
 import io.spine.validation.ConstraintViolation
-import io.spine.validation.ErrorPlaceholder
-import io.spine.validation.ErrorPlaceholder.FIELD_DUPLICATES
-import io.spine.validation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.Placeholder
+import io.spine.validation.StandardPlaceholder.FIELD_DUPLICATES
+import io.spine.validation.StandardPlaceholder.FIELD_PATH
+import io.spine.validation.StandardPlaceholder.FIELD_TYPE
+import io.spine.validation.StandardPlaceholder.FIELD_VALUE
+import io.spine.validation.StandardPlaceholder.PARENT_TYPE
 
 /**
  * The generator for the `(distinct)` option.
@@ -179,7 +179,7 @@ private class GenerateDistinct(private val view: DistinctField) {
         typeName: Expression<String>,
         fieldValue: Expression<*>,
         duplicates: Expression<*>
-    ): Map<ErrorPlaceholder, Expression<String>> = mapOf(
+    ): Map<Placeholder, Expression<String>> = mapOf(
         FIELD_PATH to fieldPath.joinToString(),
         FIELD_VALUE to fieldType.stringValueOf(fieldValue),
         FIELD_TYPE to StringLiteral(fieldType.name),
