+++ java/src/main/kotlin/io/spine/tools/validation/java/generate/option/RequiredGenerator.kt
@@ -50,10 +50,10 @@ import io.spine.tools.validation.java.generate.ValidateScope.parentPath
 import io.spine.tools.validation.java.generate.ValidateScope.violations
 import io.spine.tools.validation.option.IF_MISSING
 import io.spine.validation.ConstraintViolation
-import io.spine.validation.ErrorPlaceholder
-import io.spine.validation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.Placeholder
+import io.spine.validation.StandardPlaceholder.FIELD_PATH
+import io.spine.validation.StandardPlaceholder.FIELD_TYPE
+import io.spine.validation.StandardPlaceholder.PARENT_TYPE
 
 /**
  * The generator for `(required)` option.
@@ -116,7 +116,7 @@ private class GenerateRequired(
     private fun supportedPlaceholders(
         fieldPath: Expression<FieldPath>,
         typeName: Expression<String>,
-    ): Map<ErrorPlaceholder, Expression<String>> = mapOf(
+    ): Map<Placeholder, Expression<String>> = mapOf(
         FIELD_PATH to fieldPath.joinToString(),
         FIELD_TYPE to StringLiteral(field.type.name),
         PARENT_TYPE to typeName
