+++ java/src/main/kotlin/io/spine/tools/validation/java/setonce/SetOnceConstraintViolation.kt
@@ -38,11 +38,12 @@ import io.spine.tools.validation.java.expression.constraintViolation
 import io.spine.tools.validation.java.expression.templateString
 import io.spine.tools.validation.option.IF_SET_AGAIN
 import io.spine.validation.ConstraintViolation
-import io.spine.validation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.ErrorPlaceholder.FIELD_PROPOSED_VALUE
-import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.Placeholder
+import io.spine.validation.StandardPlaceholder.FIELD_PATH
+import io.spine.validation.StandardPlaceholder.FIELD_PROPOSED_VALUE
+import io.spine.validation.StandardPlaceholder.FIELD_TYPE
+import io.spine.validation.StandardPlaceholder.FIELD_VALUE
+import io.spine.validation.StandardPlaceholder.PARENT_TYPE
 
 /**
  * Builds a [ConstraintViolation] instance for the given field.
@@ -88,7 +89,7 @@ internal class SetOnceConstraintViolation(
     private fun supportedPlaceholders(
         currentValue: Expression<String>,
         newValue: Expression<String>
-    ) = mapOf(
+    ): Map<Placeholder, Expression<String>> = mapOf(
         FIELD_PATH to StringLiteral(fieldName),
         FIELD_TYPE to StringLiteral(fieldType),
         FIELD_VALUE to currentValue,
