+++ jvm-runtime/src/main/kotlin/io/spine/validation/TemplateStringExts.kt
@@ -29,11 +29,11 @@
 package io.spine.validation
 
 import io.spine.code.proto.FieldDeclaration
-import io.spine.validation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.ErrorPlaceholder.GOES_COMPANION
-import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
-import io.spine.validation.ErrorPlaceholder.REGEX_PATTERN
+import io.spine.validation.StandardPlaceholder.FIELD_PATH
+import io.spine.validation.StandardPlaceholder.FIELD_TYPE
+import io.spine.validation.StandardPlaceholder.GOES_COMPANION
+import io.spine.validation.StandardPlaceholder.PARENT_TYPE
+import io.spine.validation.StandardPlaceholder.REGEX_PATTERN
 import kotlin.collections.iterator
 
 /**
