+++ jvm-runtime/src/main/kotlin/io/spine/validation/TimestampValidator.kt
@@ -32,8 +32,8 @@ import com.google.protobuf.util.Timestamps
 import com.google.protobuf.util.Timestamps.MAX_VALUE
 import com.google.protobuf.util.Timestamps.MIN_VALUE
 import io.spine.base.fieldPath
-import io.spine.validation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.ErrorPlaceholder.RANGE_VALUE
+import io.spine.validation.StandardPlaceholder.FIELD_PATH
+import io.spine.validation.StandardPlaceholder.RANGE_VALUE
 
 /**
  * Validates [Timestamp] messages.
